MonitoreosDeSeguridad/requirements.txt
sqlalchemy==1.4.40
mysqlclient==2.1.1
pandas==1.5.1
openpyxl==3.0.10
XlsxWriter
